+++ packages/client/src/components/dispatch/active-units/MergeUnitModal.tsx
@@ -23,9 +23,9 @@ interface Props {
 }
 
 export function MergeUnitModal({ unit, isDispatch, onClose }: Props) {
-  const { activeOfficer } = useLeoState();
+  const { activeOfficer, setActiveOfficer } = useLeoState();
   const { isOpen, closeModal } = useModal();
-  const { activeOfficers } = useActiveOfficers();
+  const { activeOfficers, setActiveOfficers } = useActiveOfficers();
   const { state, execute } = useFetch();
   const common = useTranslations("Common");
   const t = useTranslations("Leo");
@@ -54,6 +54,21 @@ export function MergeUnitModal({ unit, isDispatch, onClose }: Props) {
     });
 
     if (json.id) {
+      const newOfficers = [];
+
+      for (const officer of activeOfficers) {
+        if (values.ids.some((v) => v.value === officer.id)) {
+          continue;
+        }
+
+        newOfficers.push(officer);
+      }
+
+      if (!isDispatch) {
+        setActiveOfficer(json);
+      }
+
+      setActiveOfficers([json, ...newOfficers]);
       handleClose();
     }
   }
