+++ packages/client/src/components/shared/NotepadModal.tsx
