+++ packages/client/src/pages/dispatch/index.tsx
@@ -25,7 +25,7 @@ import { useModal } from "state/modalState";
 import { Permissions } from "@snailycad/permissions";
 
 const NotepadModal = dynamic(async () => {
-  return (await import("components/modals/NotepadModal")).NotepadModal;
+  return (await import("components/shared/NotepadModal")).NotepadModal;
 });
 
 const WeaponSearchModal = dynamic(async () => {
