+++ packages/client/src/pages/officer/index.tsx
@@ -24,7 +24,7 @@ import { ModalIds } from "types/ModalIds";
 import { Permissions } from "@snailycad/permissions";
 
 const NotepadModal = dynamic(async () => {
-  return (await import("components/modals/NotepadModal")).NotepadModal;
+  return (await import("components/shared/NotepadModal")).NotepadModal;
 });
 
 const SelectOfficerModal = dynamic(async () => {
