+++ packages/utils/src/callsign.ts
@@ -11,10 +11,6 @@ type Unit = Pick<Officer, P | "divisions"> | Pick<EmsFdDeputy, P | "division"> |
  */
 export function generateCallsign(unit: Unit, template: string | null) {
   const isCombined = !("citizenId" in unit) || "officers" in unit;
-
-  const callsign = unit.callsign;
-  const callsign2 = unit.callsign2;
-  const department = unit.department;
   const incremental = isCombined ? unit.incremental : null;
 
   const unitDivision =
@@ -26,9 +22,9 @@ export function generateCallsign(unit: Unit, template: string | null) {
   }
 
   const replacers = {
-    department: department?.callsign,
-    callsign1: callsign,
-    callsign2,
+    department: unit.department?.callsign,
+    callsign1: unit.callsign,
+    callsign2: unit.callsign2,
     division: division?.callsign,
     incremental,
   };
